24/05 16:05 - fix(position): ripristino CRUD + controllo eliminazione posizione + supporto warning contestuale

+++ app/src/main/java/com/example/boxmanagernew/data/repository/LocationRepositoryImpl.kt
@@ -90,14 +90,23 @@ class LocationRepositoryImpl(
         location: Location
     ): Boolean {
 
-        val entity =
+        val usedCount =
+            boxDao.countBoxesByPosition(
+                location.name
+            )
+
+        if (
+            usedCount > 0
+        ) {
+
+            return false
+        }
+
+        locationDao.delete(
             LocationEntity(
                 id = location.id,
                 name = location.name
             )
-
-        locationDao.delete(
-            entity
         )
 
         return true
